Moved LIPS to metaheuristics package.
